+++ test/unit/AssetManagedPair.t.sol
@@ -90,13 +90,13 @@ contract AssetManagedPairTest is BaseTest {
         address lToken1 = _pair.token1();
 
         // sanity
-        (uint112 lReserve0, uint112 lReserve1,) = _pair.getReserves();
+        (uint104 lReserve0, uint104 lReserve1,) = _pair.getReserves();
         uint256 lBal0Before = IERC20(lToken0).balanceOf(address(_pair));
         uint256 lBal1Before = IERC20(lToken1).balanceOf(address(_pair));
 
         _manager.adjustManagement(_pair, 20e18, 20e18);
 
-        (uint112 lReserve0_1, uint112 lReserve1_1,) = _pair.getReserves();
+        (uint104 lReserve0_1, uint104 lReserve1_1,) = _pair.getReserves();
         uint256 lBal0After = IERC20(lToken0).balanceOf(address(_pair));
         uint256 lBal1After = IERC20(lToken1).balanceOf(address(_pair));
 
@@ -113,7 +113,7 @@ contract AssetManagedPairTest is BaseTest {
         // act
         _manager.adjustManagement(_pair, lAmount0Decrease, lAmount1Decrease);
 
-        (uint112 lReserve0_2, uint112 lReserve1_2,) = _pair.getReserves();
+        (uint104 lReserve0_2, uint104 lReserve1_2,) = _pair.getReserves();
 
         // assert
         assertEq(uint256(lReserve0_2), lReserve0);
@@ -126,7 +126,7 @@ contract AssetManagedPairTest is BaseTest {
 
     function testAdjustManagement_KStillHolds(uint256 aMintAmt) external allPairs {
         // assume
-        uint256 lMintAmt = bound(aMintAmt, 1, type(uint112).max / 3);
+        uint256 lMintAmt = bound(aMintAmt, 1, type(uint104).max / 3);
 
         // arrange
         vm.prank(address(_factory));
@@ -157,7 +157,7 @@ contract AssetManagedPairTest is BaseTest {
         _pair.setManager(_manager);
 
         _manager.adjustManagement(_pair, 10e18, 10e18);
-        _manager.adjustBalance(_pair, address(_tokenA), uint112(lNewManagedBalance0)); // some amount lost
+        _manager.adjustBalance(_pair, address(_tokenA), uint104(lNewManagedBalance0)); // some amount lost
 
         // sanity
         uint256 lTokenAManaged = _manager.getBalance(_pair, address(_tokenA));
@@ -184,8 +184,8 @@ contract AssetManagedPairTest is BaseTest {
         _pair.setManager(_manager);
 
         _manager.adjustManagement(_pair, 10e18, 10e18);
-        _manager.adjustBalance(_pair, address(_tokenA), uint112(lNewManagedBalance0)); // some amount lost
-        _manager.adjustBalance(_pair, address(_tokenB), uint112(lNewManagedBalance1)); // some amount lost
+        _manager.adjustBalance(_pair, address(_tokenA), uint104(lNewManagedBalance0)); // some amount lost
+        _manager.adjustBalance(_pair, address(_tokenB), uint104(lNewManagedBalance1)); // some amount lost
 
         // act
         _tokenA.mint(address(_pair), 100e18);
@@ -210,8 +210,8 @@ contract AssetManagedPairTest is BaseTest {
         _pair.setManager(_manager);
 
         _manager.adjustManagement(_pair, 10e18, 10e18);
-        _manager.adjustBalance(_pair, address(_tokenA), uint112(lNewManagedBalance0)); // some amount lost
-        _manager.adjustBalance(_pair, address(_tokenB), uint112(lNewManagedBalance1)); // some amount lost
+        _manager.adjustBalance(_pair, address(_tokenA), uint104(lNewManagedBalance0)); // some amount lost
+        _manager.adjustBalance(_pair, address(_tokenB), uint104(lNewManagedBalance1)); // some amount lost
 
         // act
         uint256 lLpTokenBal = _pair.balanceOf(_alice);
@@ -241,7 +241,7 @@ contract AssetManagedPairTest is BaseTest {
         _pair.setManager(_manager);
 
         _manager.adjustManagement(_pair, 10e18, 10e18);
-        _manager.adjustBalance(_pair, address(_tokenA), uint112(lNewManagedBalance0)); // some amount lost
+        _manager.adjustBalance(_pair, address(_tokenA), uint104(lNewManagedBalance0)); // some amount lost
 
         _pair.sync();
 
@@ -280,8 +280,8 @@ contract AssetManagedPairTest is BaseTest {
         assertEq(_manager.getBalance(_constantProductPair, lToken1), 20e18);
 
         // act
-        _manager.adjustBalance(_constantProductPair, lToken0, uint112(lNewManagedBalance0)); // some amount lost
-        _manager.adjustBalance(_constantProductPair, lToken1, uint112(lNewManagedBalance1)); // some amount lost
+        _manager.adjustBalance(_constantProductPair, lToken0, uint104(lNewManagedBalance0)); // some amount lost
+        _manager.adjustBalance(_constantProductPair, lToken1, uint104(lNewManagedBalance1)); // some amount lost
         _constantProductPair.transfer(address(_constantProductPair), 10e18);
         _constantProductPair.burn(address(this));
 
@@ -317,15 +317,15 @@ contract AssetManagedPairTest is BaseTest {
         assertEq(_manager.getBalance(_stablePair, lToken1), 20e18);
 
         // act
-        _manager.adjustBalance(_stablePair, lToken0, uint112(lNewManagedBalance0)); // some amount lost
-        _manager.adjustBalance(_stablePair, lToken1, uint112(lNewManagedBalance1)); // some amount lost
+        _manager.adjustBalance(_stablePair, lToken0, uint104(lNewManagedBalance0)); // some amount lost
+        _manager.adjustBalance(_stablePair, lToken1, uint104(lNewManagedBalance1)); // some amount lost
         _stablePair.transfer(address(_stablePair), 10e18);
         _stablePair.burn(address(this));
 
         // assert
         assertEq(_manager.getBalance(_stablePair, lToken0), lNewManagedBalance0);
         assertEq(_manager.getBalance(_stablePair, lToken1), lNewManagedBalance1);
-        (uint112 lReserve0, uint112 lReserve1,) = _stablePair.getReserves();
+        (uint104 lReserve0, uint104 lReserve1,) = _stablePair.getReserves();
         assertTrue(
             MathUtils.within1(lReserve0, (INITIAL_MINT_AMOUNT - 20e18 + lNewManagedBalance0 + 10e18) * 210e18 / 220e18)
         );
@@ -340,8 +340,8 @@ contract AssetManagedPairTest is BaseTest {
         // assume
         int256 lAmount0Managed = int256(bound(aAmount0, 1, INITIAL_MINT_AMOUNT));
         int256 lAmount1Managed = int256(bound(aAmount1, 1, INITIAL_MINT_AMOUNT));
-        uint112 lAmount0NewBalance = uint112(bound(aNewAmount0, 1, type(uint112).max / 2));
-        uint112 lAmount1NewBalance = uint112(bound(aNewAmount1, 1, type(uint112).max / 2));
+        uint104 lAmount0NewBalance = uint104(bound(aNewAmount0, 1, type(uint104).max / 2));
+        uint104 lAmount1NewBalance = uint104(bound(aNewAmount1, 1, type(uint104).max / 2));
 
         // arrange
         vm.prank(address(_factory));
@@ -354,7 +354,7 @@ contract AssetManagedPairTest is BaseTest {
         _pair.sync();
 
         // assert
-        (uint112 lReserve0, uint112 lReserve1,) = _pair.getReserves();
+        (uint104 lReserve0, uint104 lReserve1,) = _pair.getReserves();
         assertEq(_pair.token0Managed(), lAmount0NewBalance);
         assertEq(_pair.token1Managed(), lAmount1NewBalance);
         assertEq(lReserve0, INITIAL_MINT_AMOUNT - uint256(lAmount0Managed) + lAmount0NewBalance);
